Rewrite template and design a simple layout
I implement a simple design for now. I added coming soon poster for some features.
After my class, I will try to plan what features to add for proof of the concept.

This project aims to provide the ability to build immutable lockers from
different types. And, provide the ability to distribute them with a proof of
existence if needed.

- Use Node v16 for the sandbox
- Design a simple layout for pages
- Rewrite the main page