Implement the first builder for immutable lockers! \o/
This is the first version (aka v1) for immutable locker protocol.
This protocol currently supports encrypting messages for now only.

The metadata is unencrypted and open to everyone.
Note: I'm thinking of providing two-way encryption for the metadata.
Because, it doesn't have to be public to everyone.

This version provides a strong way of generating an encryption key
from a passphrase no matter how weak it is.

Implementation:

The passphrase is being hashed with SHA-256.
(Note: The passphrase is normalized with `NFKC` and encoded to bytes)
Then, we generate a 64-bit of salt (CSPRN). Then, we generate a
password-based key derivation using "scrypt" to provide the locker
the ability for the locker key to be encrypted forever.
(SCRYPT PARAMS: N=2^16, R=16, P=1)

After generating the encryption key we can use this key to
encrypt the locker data with AES-256.

Important Note: The locker only stores the salt for uniqueness.

the file with ".locker.checksum" is used to verify the given locker!