Implement a locker viewer + re-organizing scripts
The viewer is very simple and not stylish at all.
It takes the locker data and tries to unlock it
or basically shows what you can view on it.

Some lockers with missing verification messages
(like example/v1.locker) doesn't have a way to verify
if the content shown is correct or not. So, it shows
the content of the locker whether it was encrypted or
not.