Implement tool to verify an immutable locker
This tools helps the user verify an immutable locker using SHA-512
and compare it with another SHA-512. The target SHA-512 can be
read as HEX. It converts the HEX format to binary in order to compare
it with another binary format. This is efficient and useful for verifying
the existence of an immutable locker on a blockchain network, or through
a message.